# src/main/resources/static/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Lunit service APIs",
        "description": "Test project for Lunit Assignment",
        "contact": {
            "name": "Bulgan Galbadrakh",
            "email": "bulgan.galbadrakh@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local ENV"
        }
    ],
    "paths": {
        "/users/{userId}/slides/getAnalysis": {
            "put": {
                "tags": [
                    "slide-controller"
                ],
                "summary": "Send request to do analysis on all selected slides from AI client",
                "operationId": "sendRequestToAIClient",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Slide"
                                }
                            },
                            "example": [
                                {
                                    "id": 4,
                                    "name": "CMU-1-Small-Region.svs",
                                    "isProcessed": false,
                                    "status": "NONE",
                                    "decision": null,
                                    "score": null,
                                    "description": "Exported region from CMU-1.svs",
                                    "format": "Aperio SVS",
                                    "license": "CC0-1.0",
                                    "size": 1938955,
                                    "sha256": "ed92d5a9f2e86df6"
                                }
                            ]
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{userId}/slides/upload": {
            "post": {
                "tags": [
                    "slide-controller"
                ],
                "summary": "Upload a new slide and store in DB",
                "operationId": "uploadSlides",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Slide"
                                }
                            },
                            "example": [
                                {
                                    "name": "CMU-1-Small-Region.svs",
                                    "isProcessed": false,
                                    "status": "NONE",
                                    "decision": null,
                                    "score": null,
                                    "description": "Exported region from CMU-1.svs",
                                    "format": "Aperio SVS",
                                    "license": "CC0-1.0",
                                    "size": 1938955,
                                    "sha256": "ed92d5a9f2e86df6"
                                }
                            ]
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Slide"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/users/{userId}/slides": {
            "get": {
                "tags": [
                    "slide-controller"
                ],
                "summary": "Get all existing slides of one user",
                "operationId": "getAllSlidesOfUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "slideId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{userId}/slides/requestHistory": {
            "get": {
                "tags": [
                    "slide-controller"
                ],
                "summary": "Get a list of all requested slides by userId",
                "operationId": "getAllRequestedSlidesOfUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "slideId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Slide": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "isProcessed": {
                        "type": "boolean"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "NONE",
                            "PENDING",
                            "COMPLETED",
                            "FAILED"
                        ]
                    },
                    "decision": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "score": {
                        "type": "number",
                        "format": "double"
                    },
                    "description": {
                        "type": "string"
                    },
                    "format": {
                        "type": "string"
                    },
                    "license": {
                        "type": "string"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "sha256": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
        }
    }
}